scripts/func/requirements.txt
azure-functions
youtube_transcript_api